# Small performance improvement for `BinaryExpression`